Builded a new simple game for control
I created a simple space game for optimizing the movement of ball and testing object detection control system with new simple sphare 3D Object called "Ball",  3D ring prefab called "RingObs", and with Space SkyBox. Main Ball (it's just simple primitive sphare, can be create into hierarchy) is getting the coordinates of gray ball and fallows it smootly with a minimized delay. mainBall.cs should be in Player  object, gameControl.cs should be in Ball,  ringObs.cs should be in ring prefab and spawnRing.cs should be in "Ring Spawner" game object that positioned desired coordinate in the unity world space . 
And all mesh behaviours in player should be removed. The purpose of game is passing the Ball through the ring obs. That's it.